+++ Excel_PRIME.RangeBench/GetRangeAsposeCells.cs
@@ -23,19 +23,15 @@ public class GRAsposeCells : IGetRange
         return true;
     }
 
-    public IEnumerable<IEnumerable<object?>> GetDefinedRange(string definedName, int? localSheetId = null)
+    public IEnumerable<IEnumerable<object?>> GetDefinedRange(string definedName, string? sheetName = null)
     {
         Aspose.Cells.Range namedRange;
-        if (localSheetId.HasValue)
-        {
+        
+        namedRange = !string.IsNullOrEmpty(sheetName)
             //Get a specific named range in the worksheet
-            namedRange = _workbook.Worksheets.GetRangeByName(definedName, localSheetId.Value, false);
-        }
-        else
-        {
+            ? null! //_workbook.Worksheets.GetRangeByName(definedName, sheetName, false)
             //Get a specific named range in the workbook
-            namedRange = _workbook.Worksheets.GetRangeByName(definedName);
-        }
+            : _workbook.Worksheets.GetRangeByName(definedName);
         if (namedRange != null)
         {
             //Get range