+++ Excel_PRIME/IXmlReaderHelpers.cs
@@ -70,9 +70,4 @@ public interface IOpenXmlReaderHelpersAsync : IOpenXmlReaderHelpers
     /// <returns></returns>
     Task<IOpenXmlSheetReaderAsync> CreateSheetReaderAsync(NonClosingStream stream, InstanceContext instanceContext,
         XmlNameTable sharedNameTable, CancellationToken ct);
-
-    /// <summary>
-    /// Get the internal file name of this worksheet type
-    /// </summary>
-    string GetSheetFileName(int offsetSheetId);
 }